Add files via upload
Python script to automate ROPS-type analyses in Abaqus/Standard

A ROPS (Roll Over Protection System) type analysis consists of a series of restart jobs which involve pre-loading, loading and unloading the ROPS structure. The analysis should be terminated when the structure has absorbed a set amount of energy. In the normal process a user manually monitors this and then terminates a particular job, which is tedious and requires constant monitoring. 

 

The script attached here will run a series of restart jobs and terminate the specified jobs which satisfy an energy absorption requirement. Those jobs which do not have a energy requirement will be allowed to run to completion. For each job which has an energy requirement, the user needs to specify the requirement in the spreadsheet. The Internal Energy (ALLIE) is tracked using the data file after requesting the keyword *ENERGY PRINT in the input file. This keyword must be placed in each input file.

 

The script does not contain any Abaqus/CAE modules, so it only consumes Abaqus/Standard tokens.

 

The script is demonstrated with a simple model for two loading and unloading cycles → FRAME-LOADING-1 → FRAME-LOADING-2 → FRAME-LOADING-3 → FRAME-LOADING-4.

 

Please note that to terminate an analysis accurately at required energy, the maximum time increment must be properly specified for the jobs with energy requirement.

 

Please delete all old job files (dat, msg etc) in the folder before running this script.
 

This script requires the file SPREADSHEET_FOR_JOBS.csv to run. Inputs such as job names, energy requirements are to be mentioned in this file.


Run the script using these commands and in the directory where all the input files have been placed:

C:\SIMULIA\Commands\abq2018 python ROPS-AUTOMATE.py
OR

abaqus python ROPS-AUTOMATE.py

OR

abq2018 python ROPS-AUTOMATE.py
 

Disclaimer

The attachments on this article are subject to certain usage conditions.  Please click here for details.